Desaclopei algumas coisas

+++ src/main/groovy/com/jp/gas_station/Models/profitPerPump.groovy
@@ -3,7 +3,7 @@ package com.jp.gas_station.Models
 class profitPerPump {
     Integer pumpId
     String type
-    Integer sellingPrice
-    Integer totalWasted
+    Double sellingPrice
+    Double totalWasted
     Double totalProfit
 }
